[CTR] / 마법의 숲 탐색 / GOLD 3 / 5.5H
